Reduced maximum delay on Effect after mouseup
Gradient seemed to linger too long (especially on the dark demo image
and large header). Also reduces noticability of jagged gradient on the
big header in Chrome.

In FF the header animation seems to stutter, should we revert back to
the default version if the `div` with the ripple effect is too large, or
is this overthinking?

I probably shouldn't post suggestions inside commits...