Added .waves-classic, prettified gradients (#50)
Had to slow down the default Effect.duration (sorry!). It's too quick
for the radial gradients. In theory we could filter between
.waves-classic to set the default timing but that's probably more
complexity than it's worth.

I wonder if there's an easy way we can let the user configure the
gradient? As you can see, the radial gradient looks great on the boxes
and images, but not so great on the plain icons (hence why I added
.waves-classic).

Let me know what you think of the radial gradients though; and maybe
this answers Improvement #50 slightly (letting the user set their own
gradient and timing?).